feat(js): added useEffect for utilityFetchReviews() and moved everything to MoviePage.jsx
