+++ src/Context.jsx
@@ -35,22 +35,6 @@ const AppProvider = ({ children }) => {
 		}
 	}
 
-	async function utilityFetchReviews() {
-		try {
-			const response = await fetch("http://localhost:8080/films/:idFilm/reviews");
-
-			if (!response.ok) {
-				throw new Error("Network response was not ok");
-			}
-
-			const data = await response.json();
-			localStorage.setItem("reviews", JSON.stringify(data));
-			return data;
-		} catch (error) {
-			console.error("Error fetching reviews:", error);
-			throw error;
-		}
-	}
 
 	useEffect(() => {
 		if (!movies.length) {
